+++ tests/test_torrent.py
@@ -1,3 +1,4 @@
+import maya
 import pytest
 
 from .context import mariner
@@ -11,7 +12,7 @@ class TestTorrent:
 
     @pytest.fixture
     def torrent(self):
-        return torrent.Torrent('name', 'tracker', magnet='magnet', torrent='tracker')
+        return torrent.Torrent('name', 'tracker', date='yesterday', magnet='magnet', torrent='tracker')
 
     def test_name_is_string(self, torrent):
         """Name is a string."""
@@ -36,3 +37,7 @@ class TestTorrent:
     def test_cmpkey_is_int(self, torrent):
         """_cmpkey is an integer."""
         assert isinstance(torrent._cmpkey, int)
+
+    def test_date_is_mayadt(self, torrent):
+        """Filename is a string."""
+        assert isinstance(torrent.date, maya.core.MayaDT)