+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Sort results by upload date using --newest/-n flag.
 - Search all available trackers using --all/-a option.
 - Colored output
+- Docker image
+- Support snaps
 
 ### Changed
 - Switch default tracker to Linuxtracker