+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [1.0.0] - 2018-02-28
+## [1.0.1] - 2018-02-28
 ### Fixed
 - Restore compatibility with Python 3.5
 
@@ -19,7 +19,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Changed
 - Switch default tracker to Linuxtracker.
-
 
 ## [0.4.0] - 2018-02-01
 ### Added